-Added shape flip -Added colored shapes -Added scoring and leveling system -Fixed some minor bugs

+++ include/grid.h
@@ -47,7 +47,7 @@ struct Grid
     int yCells;
 
     // Cells boder thickness and color
-    unsigned int cellsBorder;
+    float cellsBorder;
     SDL_Color cellsBorderColor;
 
     // Matrix of Cells
